+++ packages/retail-ui/components/DateInput/DateInputHelpers/UnknownDatePart.tsx
@@ -1,5 +1,0 @@
-export class UnknownDatePart extends Error {
-  constructor() {
-    super('Unknown date part');
-  }
-}