+++ packages/retail-ui/components/LocaleProvider/decorators.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
+import { defaultLangCode } from './constants';
 import { LocaleConsumer, LocaleProviderProps } from './LocaleProvider';
 import { LocaleHelper } from './LocaleHelper';
-import { LocaleControls } from './types';
+import { LangCodes, LocaleControls } from './types';
 
 export function locale<C>(controlName: keyof LocaleControls, localeHelper: LocaleHelper<C>) {
   return <T extends new (...args: any[]) => React.Component>(constructor: T) => {
@@ -23,9 +24,16 @@ export function locale<C>(controlName: keyof LocaleControls, localeHelper: Local
       }
 
       public get locale(): C {
+        if (this._localeContext === undefined) {
+          return {} as C;
+        }
         const langCode = this._localeContext.langCode;
         const localeFromContext = this._localeContext.locale && this._localeContext.locale[this.controlName];
         return Object.assign({}, this.localeHelper.get(langCode), localeFromContext);
+      }
+
+      public get langCode(): LangCodes {
+        return this._localeContext.langCode || defaultLangCode;
       }
     };
     Object.defineProperty(LocaleDecorator, 'name', { value: constructor.name });